backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
openai==1.47.0
python-docx==1.1.2
httpx==0.27.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
stripe==10.10.0
redis==5.0.8
celery==5.4.0
boto3==1.35.35
alembic==1.13.2
python-json-logger==2.0.7
cryptography==43.0.1
python-dotenv==1.0.1
pytest==8.3.3
